refactor(#1470): decompose engine/scheduler.py per ADR-046 (Phase 3 D1)
Split the 1744-LOC ``src/scistudio/engine/scheduler.py`` god-file into
the ``src/scistudio/engine/scheduler/`` sub-package per the Path D
pattern from ADR-046 (Phase 3 D1 of umbrella #1427). The
``DAGScheduler`` class and ``RunHandle`` dataclass stay defined in
``__init__.py``; method bodies live as private ``_underscore``
module-level functions in five sibling modules and bind onto
``DAGScheduler`` via class-body static assignment.

Sub-package layout (all under the 750 LOC god-file threshold):

* ``__init__.py``  (521 LOC) — class + bindings + helper re-exports
* ``_dispatch.py`` (570 LOC) — block dispatch loop
* ``_events.py``   (274 LOC) — EventBus subscriber handlers
* ``_lineage.py``  (367 LOC) — output + lineage persistence
* ``_helpers.py``  ( 68 LOC) — module-level helpers re-imported by __init__
* ``_state.py``    ( 73 LOC) — state-machine helpers (#1449 contract)
* ``_rerun.py``    (169 LOC) — rerun / reset / graph-traversal

ADR-046 §5 only-move: ``_check_readiness``, ``_check_completion``,
``_propagate_skip`` move to ``_state.py`` byte-identically. The #1449
scheduler state-machine contract test passes unedited.

§C9 compliance: ``_dispatch.py`` / ``_events.py`` / ``_lineage.py`` /
``_state.py`` / ``_rerun.py`` / ``_helpers.py`` contain zero
``ClassDef`` nodes. The new
``tests/engine/test_scheduler_package_layout.py`` AST walker pins this
contract.

Public import surface preserved: ``DAGScheduler``, ``RunHandle``,
``logger``, ``_MAX_ERROR_SUMMARY_LEN``, ``_extract_error_summary``,
``_collect_object_ids``, ``_object_ids_for_value`` all remain
importable from ``scistudio.engine.scheduler``.

Frontmatter ``governs.files`` updated in ADR-046 (expanded to the
new sub-package), ADR-018, ADR-023, ADR-044 (each previously pointed
at the deleted ``scheduler.py``). ``scripts/check_god_files.py``
waiver entry removed.

Closes #1470

Gate-Record: .workflow/records/1470-scheduler-decomp.json
Task-Kind: refactor
Issue: #1470
