+++ .workflow/records/1470-scheduler-decomp.json
@@ -1,0 +1,196 @@
+{
+  "admin_labels": [
+    {
+      "applied_at": null,
+      "applied_by": null,
+      "name": "admin-approved:core-change"
+    }
+  ],
+  "amendments": [],
+  "branch": "refactor/issue-1470/scheduler-decomp",
+  "changed_test_paths": [
+    "tests/engine/test_scheduler_package_layout.py"
+  ],
+  "check_results": [
+    {
+      "command_or_tool": "python -m ruff check src/scistudio/engine/scheduler/ tests/engine/test_scheduler_package_layout.py scripts/check_god_files.py",
+      "exit_code": 0,
+      "name": "ruff",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "python -m ruff format --check src/scistudio/engine/scheduler/ tests/engine/test_scheduler_package_layout.py scripts/check_god_files.py",
+      "exit_code": 0,
+      "name": "format",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "python -m pytest tests/engine/ tests/integration/test_cancel_scenario.py tests/integration/test_multimodal_workflow.py tests/integration/test_smoke_workflows.py tests/contracts/test_eventbus_lifecycle_contract.py tests/contracts/test_lineage_run_recipe_contract.py tests/core/test_lineage_store_4table.py --timeout=60 -x --no-cov",
+      "exit_code": 0,
+      "name": "pytest",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    },
+    {
+      "command_or_tool": "python scripts/check_god_files.py",
+      "exit_code": 0,
+      "name": "god_file_advisory",
+      "output_path": null,
+      "status": "pass",
+      "summary": {},
+      "timestamp": null
+    }
+  ],
+  "commit": null,
+  "docs_landing": {
+    "adr": {
+      "paths": [
+        "docs/adr/ADR-046.md",
+        "docs/adr/ADR-018.md",
+        "docs/adr/ADR-023.md",
+        "docs/adr/ADR-044.md"
+      ]
+    },
+    "changelog": {
+      "not_applicable": true,
+      "rationale": "Structural refactor only; zero user-visible change. Phase 3 D1 of umbrella 1427."
+    },
+    "checklist": {
+      "not_applicable": true,
+      "rationale": "Single-PR implementation; no multi-agent dispatch."
+    },
+    "docs": {
+      "paths": [
+        "docs/adr/ADR-046.md",
+        "docs/adr/ADR-018.md",
+        "docs/adr/ADR-023.md",
+        "docs/adr/ADR-044.md"
+      ]
+    },
+    "spec": {
+      "not_applicable": true,
+      "rationale": "ADR-046 authoritative; no separate spec."
+    }
+  },
+  "full_audit": {
+    "blocks_merge": false,
+    "command": "python -m scistudio.qa.audit.full_audit",
+    "exit_code": 0,
+    "known_debt": [],
+    "output_path": "docs/audit/full-audit-latest.json",
+    "status": "pass",
+    "summary": {},
+    "unclassified_failures": []
+  },
+  "governance_touch": false,
+  "issues": [
+    {
+      "close_in_pr": true,
+      "followup_rationale": null,
+      "number": 1470,
+      "url": "https://github.com/zjzcpj/SciStudio/issues/1470"
+    }
+  ],
+  "owner_directive": "Implement ADR-046 scheduler decomposition",
+  "planned_files": [
+    "src/scistudio/engine/scheduler/__init__.py",
+    "src/scistudio/engine/scheduler/_dispatch.py",
+    "src/scistudio/engine/scheduler/_events.py",
+    "src/scistudio/engine/scheduler/_helpers.py",
+    "src/scistudio/engine/scheduler/_lineage.py",
+    "src/scistudio/engine/scheduler/_rerun.py",
+    "src/scistudio/engine/scheduler/_state.py",
+    "tests/engine/test_scheduler_package_layout.py",
+    "scripts/check_god_files.py",
+    "docs/adr/ADR-046.md",
+    "docs/adr/ADR-018.md",
+    "docs/adr/ADR-023.md",
+    "docs/adr/ADR-044.md"
+  ],
+  "pull_request": null,
+  "record_path": ".workflow/records/1470-scheduler-decomp.json",
+  "required_checks": [
+    "ruff",
+    "format",
+    "pytest",
+    "god_file_advisory",
+    "full_audit",
+    "sentrux"
+  ],
+  "schema_version": "1",
+  "scope": {
+    "exclude": [],
+    "include": [
+      "src/scistudio/engine/scheduler/**",
+      "src/scistudio/engine/scheduler.py",
+      "tests/engine/test_scheduler_package_layout.py",
+      "scripts/check_god_files.py",
+      "docs/adr/ADR-046.md",
+      "docs/adr/ADR-018.md",
+      "docs/adr/ADR-023.md",
+      "docs/adr/ADR-044.md",
+      ".workflow/records/1470-scheduler-decomp.json"
+    ]
+  },
+  "sentrux": {
+    "command_or_tool": "mcp__plugin_sentrux_sentrux__scan + check_rules + health",
+    "mode": "free-tier",
+    "name": "sentrux.free_tier",
+    "output_path": null,
+    "pro_required": false,
+    "quality_signal": 4233.0,
+    "rules_checked": 3,
+    "status": "pass",
+    "summary": {},
+    "thresholds": {},
+    "total_rules_defined": 15
+  },
+  "stages": [
+    {
+      "completed_at": null,
+      "stage": "scope_and_issue",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "plan",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": "2026-05-22T23:29:20.600066Z",
+      "stage": "implement",
+      "status": "done",
+      "summary": "Scheduler split into sub-package per ADR-046; docs (ADR-046/018/023/044) updated."
+    },
+    {
+      "completed_at": "2026-05-22T23:29:20.600066Z",
+      "stage": "update_docs",
+      "status": "done",
+      "summary": "Scheduler split into sub-package per ADR-046; docs (ADR-046/018/023/044) updated."
+    },
+    {
+      "completed_at": null,
+      "stage": "test_and_checks",
+      "status": "done",
+      "summary": null
+    },
+    {
+      "completed_at": null,
+      "stage": "commit_and_submit_pr",
+      "status": "pending",
+      "summary": null
+    }
+  ],
+  "task_id": "1470-scheduler-decomp",
+  "task_kind": "refactor"
+}
