+++ docs/adr/ADR-018.md
@@ -31,7 +31,7 @@ governs:
     - src/scistudio/blocks/base/state.py
     - src/scistudio/blocks/base/block.py
     - src/scistudio/engine/events.py
-    - src/scistudio/engine/scheduler.py
+    - src/scistudio/engine/scheduler/**
     - src/scistudio/engine/checkpoint.py
     - src/scistudio/core/lineage/**
     - src/scistudio/api/ws.py
