+++ docs/adr/ADR-023.md
@@ -32,7 +32,7 @@ governs:
     - src/scistudio/api/routes/workflows.py
     - src/scistudio/api/routes/projects.py
     - src/scistudio/api/routes/data.py
-    - src/scistudio/engine/scheduler.py
+    - src/scistudio/engine/scheduler/**
     - frontend/src/App.tsx
     - frontend/src/components/Toolbar.tsx
     - frontend/src/components/BlockPalette.tsx
