+++ docs/adr/ADR-044.md
@@ -33,7 +33,7 @@ governs:
     - docs/architecture/ARCHITECTURE.md
     - docs/specs/adr-044-subworkflow-block.md
   excludes:
-    - src/scistudio/engine/scheduler.py
+    - src/scistudio/engine/scheduler/**
     - src/scistudio/engine/runners/**
 
 tests:
