+++ docs/adr/ADR-046.md
@@ -22,7 +22,13 @@ governs:
   entry_points: []
   files:
     - docs/adr/ADR-046.md
-    - src/scistudio/engine/scheduler.py
+    - src/scistudio/engine/scheduler/__init__.py
+    - src/scistudio/engine/scheduler/_dispatch.py
+    - src/scistudio/engine/scheduler/_events.py
+    - src/scistudio/engine/scheduler/_helpers.py
+    - src/scistudio/engine/scheduler/_lineage.py
+    - src/scistudio/engine/scheduler/_rerun.py
+    - src/scistudio/engine/scheduler/_state.py
   excludes:
     - tests/**
     - docs/audit/**
@@ -33,6 +39,7 @@ tests:
   - tests/engine/test_scheduler_terminal_data.py
   - tests/engine/test_scheduler_rerun_cancel.py
   - tests/engine/test_scheduler_state_machine_contract.py
+  - tests/engine/test_scheduler_package_layout.py
 agent_editable: false
 assisted_by:
   - "claude:opus-4-7"
