+++ scripts/check_god_files.py
@@ -38,7 +38,10 @@ GOD_FILE_SIZE_WAIVERS: frozenset[str] = frozenset(
     {
         # ``src/scistudio/api/runtime.py`` removed 2026-05-22 (#1430) — the
         # 1839-LOC god-file was split into the ``runtime/`` sub-package.
-        "src/scistudio/engine/scheduler.py",
+        # ``src/scistudio/engine/scheduler.py`` removed 2026-05-22 (#1470) —
+        # the 1744-LOC god-file was split into the ``scheduler/`` sub-package
+        # per ADR-046 (Phase 3 D1 of umbrella #1427). Every sibling module
+        # measures below the 750 LOC threshold.
         "src/scistudio/blocks/registry.py",
         # ``src/scistudio/qa/governance/gate_record.py`` was decomposed into
         # the ``gate_record/`` sub-package in PR for #1433 (umbrella #1427);
